+++ CHANGELOG.md
@@ -10,6 +10,7 @@ CHANGELOG
   - Updated RxJava to 1.2.9.
 - Fixes:
   - Fixed potential crash in simple reading mode.
+  - Fixed UI issue when all bookmaks / notes are deleted.
 
 #### v1.18.2 (2017-03-17)
 - Updated dependencies: